# Conflicts:
#	ia-cli/src/commands/download.rs
#	ia-core/src/config.rs
#	ia-core/src/disk_pool.rs
#	ia-core/src/download.rs

+++ ia-core/src/disk_pool.rs
@@ -186,8 +186,10 @@ fn disk_space(path: &Path) -> Result<(u64, u64)> {
             )));
         }
 
-        let free = stat.f_bavail as u64 * stat.f_frsize;
-        let total = stat.f_blocks as u64 * stat.f_frsize;
+        #[allow(clippy::unnecessary_cast)]
+        let free = stat.f_bavail as u64 * stat.f_frsize as u64;
+        #[allow(clippy::unnecessary_cast)]
+        let total = stat.f_blocks as u64 * stat.f_frsize as u64;
         Ok((free, total))
     }
